+++ PHP/JSON-RPC/Docs/changelog.md
@@ -1,9 +1,9 @@
 ## version 0.8.1
 - fixed items last updated func
-- convertor for items omnis2api
-- convertor for items api2omnis
+- convertor for items omnis2api added
 - aggregator mode added
 - min&max for items_last_updated added
+- items function is universal now
 
 ## version 0.8.0
 - new auth method with client id